Add customizable error handlers (#322)
This patch adds a new feature called error handlers. It allows to define the error handling logic globally and per rule. It is now possible, for example, to return a JSON response for `Accept: application/json` requests and a HTTP Redirect response for requests that are coming from a user.

This also resolves several issues, as noted below:

Closes #204 
Closes #252
Closes #119