Docker: update the Docker Compose file as per changes made over time
Oathkeeper has gone through a couple of changes since the initial draft of the
docker compose file, considering these changes and the newly introduced
Dockerfile in the previous commit, make these changes to the docker-compose.yml:
1. Bump the version of the compose file to 3.
2. Remove the need for the postgres database app, since Oathkeeper no longer needs a
database.
3. Remove the need for the migration app, since we no longer need to migrate since
there is no database and the option is deprecated.
4. Use the newly defined Dockerfile 'Dockerfile-dc'.
5. We now serve both API and PROXY from the same app, so we don't need two
instances of the app.
6. Add sample config, rules and JWK files to `.docker_compose`, mount this via a
volume mount.

Signed-off-by: karthik nayak <karthik.188@gmail.com>