feat: Add new remote authorizer that uses request body and headers
The current remote_json authorizer does not have the ability to send
the request body of the request to authorize. This means this cannot
be taken into account while checking permissions.

Providing the request body as part of the JSON payload won't always
work as JSON cannot handle binary data.