feat: add new remote authorizer that uses request body and headers (#416)
This pull request implements a new authorizer that sends the original request body as body to the remote endpoint. This allows the remote endpoint to take the body into account in its decision.

The current remote_json authorizer does not have the ability to send
the request body of the request to authorize. This means this cannot
be taken into account while checking permissions.

Providing the request body as part of the JSON payload won't always
work as JSON cannot handle binary data.