fix: don't copy the decision endpoint request's Content-Length (#422)
We currently copy all original request headers send to the decission
endpoint back. This can include the Content-Length header which
describes the request body or response. Including the original
request Content-Length causes issues for the decission endpoint
client if the response body doesn't match the exact size.

This change makes sure the Content-Length doesn't get included in
the response body and adds a test to prevent future regressions.