feat: Ability to Configure Remote Authorizers to set Headers in AuthenticationSession (#717)
The remote authorizers may have useful context from user's permissions. So with this changes, custom authorizers using remote and remote_json can return some useful headers to be forward into the AuthenticationSession, meaning that these headers will be passed to upstream services.

For example, an user containing scopes/branches inside an organization profile has some level of data addressed to him. In this case, the upstream service need to know that, and "filter" the data according to his "branch_id". The permission that is given to the user (and the remote authorizers manages) has a record of the "branch_id", for the following responses will be returned as status code 200 (if granted) and containing a header like X-Branch-Id.

The upstream service receives the X-Branch-Id and does your thing.

The configuration requires to configure a list of "allowed headers" returning from remote authorizer, that will be accepted in the pipeline.

Co-authored-by: hackerman <3372410+aeneasr@users.noreply.github.com>