fix: don't pass websocket headers on to session store
details at https://github.com/ory/cloud/issues/76

tl;dr: When Oathkeeper is used to secure the initial connection request
to a websocket endpoint, that doesn't mean we also want to open a
websocket connection to the session store. With this commit, the
"Connection: Upgrade" and related headers are removed now, so the
session store receives a "normal" HTTP request.